highscores is working. Also the endless losegame bug has been fixed. <bug, interface,organisation
