+++ src/main/java/lasersharksgui/WinPane.java
@@ -1,5 +1,7 @@
 package lasersharksgui;
 
+import java.io.IOException;
+
 import javafx.scene.text.Text;
 import lasersharks.Highscores;
 import lasersharks.Position;
@@ -9,6 +11,14 @@ public class WinPane extends StandardPane {
   public WinPane() {
     
     addMidText("YOU WON!!!", TEXT_SCALE_SIZE_BIG, -230);
-    addMidText(Highscores.makeHighscoreString(), TEXT_SCALE_SIZE_SMALL, 100);
+    try {
+      Highscores.writeHighscore();
+      addMidText(Highscores.makeHighscoreString(), TEXT_SCALE_SIZE_SMALL, 100);
+    } catch (IOException e) {
+      addMidText("highscores failed to load", TEXT_SCALE_SIZE_MED, 100);
+      e.printStackTrace();
+    }
+
+    addMidText("try again?", TEXT_SCALE_SIZE_SMALL, -200);
   }
 }
